Move custom ember-engines `{{link-to}}` to a monkey patch.
Intead of registering a custom `LinkTo` component for the engine to use,
I think it is better (at least more in line with all the other things)
to simply monkey patch.  It also makes it simpler to move upstream when
we are ready...
