+++ addon/initializers/engines.js
@@ -6,6 +6,8 @@ import 'ember-engines/engine-instance-ext';
 import 'ember-engines/keywords/mount';
 import 'ember-engines/keywords/outlet';
 import 'ember-engines/router-dsl-ext';
+import 'ember-engines/link-to-component-ext';
+
 
 // TODO: Move to ensure they are ran prior to instantiating Ember.Application
 export function initialize() {
